Renamed the helpers file to avoid confusion in the future and from outside readers
