+++ beaconChain/typescript/helpers/stateTransitionHelpers.ts
